Rename lessons controller to user_lessons and simplify responses
- Rename LessonsController to UserLessonsController
- Update routes from /v1/lessons to /v1/user_lessons
- Simplify start and complete actions to return empty hash
- Update all tests to reflect new endpoint names and responses
- Update README with new API endpoint paths
