+++ app/controllers/v1/user_lessons_controller.rb
@@ -1,0 +1,17 @@
+module V1
+  class UserLessonsController < ApplicationController
+    before_action :use_lesson!
+
+    def start
+      UserLesson::FindOrCreate.(current_user, @lesson)
+
+      render json: {}
+    end
+
+    def complete
+      UserLesson::Complete.(current_user, @lesson)
+
+      render json: {}
+    end
+  end
+end
